# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Outcome Control Plane API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  /** Total monthly savings realised in USD */
  totalMonthlySavings: number;
  totalAnnualisedSavings: number;
  pendingRecommendations: number;
  executedActions: number;
  blockedActions: number;
  activeConnectors: number;
  avgTrustScore: number;
  savingsThisMonth: number;
}

export interface SavingsTrendPoint {
  month: string;
  savings: number;
  actions: number;
}

export interface ActionBreakdownItem {
  status: string;
  count: number;
  value: number;
}

export type ConnectorType = (typeof ConnectorType)[keyof typeof ConnectorType];

export const ConnectorType = {
  m365: "m365",
  aws: "aws",
  azure: "azure",
  gcp: "gcp",
  salesforce: "salesforce",
  slack: "slack",
  zoom: "zoom",
  github: "github",
} as const;

export type ConnectorStatus =
  (typeof ConnectorStatus)[keyof typeof ConnectorStatus];

export const ConnectorStatus = {
  connected: "connected",
  disconnected: "disconnected",
  syncing: "syncing",
  error: "error",
} as const;

export interface Connector {
  id: number;
  name: string;
  type: ConnectorType;
  status: ConnectorStatus;
  lastSync: string | null;
  recordCount: number;
  trustScore: number;
}

export type RecommendationExecutionStatus =
  (typeof RecommendationExecutionStatus)[keyof typeof RecommendationExecutionStatus];

export const RecommendationExecutionStatus = {
  AUTO_EXECUTE: "AUTO_EXECUTE",
  APPROVAL_REQUIRED: "APPROVAL_REQUIRED",
  INVESTIGATE: "INVESTIGATE",
  BLOCKED: "BLOCKED",
} as const;

export type RecommendationStatus =
  (typeof RecommendationStatus)[keyof typeof RecommendationStatus];

export const RecommendationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  executed: "executed",
} as const;

export interface Recommendation {
  id: number;
  userEmail: string;
  displayName: string;
  licenceSku: string;
  monthlyCost: number;
  annualisedCost: number;
  trustScore: number;
  executionStatus: RecommendationExecutionStatus;
  status: RecommendationStatus;
  playbook: string;
  connector: string;
  lastActivity: string | null;
  daysSinceActivity: number | null;
  rejectionReason: string | null;
  createdAt: string;
}

export type OutcomeLedgerEntryEvidence = { [key: string]: unknown };

export interface OutcomeLedgerEntry {
  id: number;
  recommendationId: number;
  userEmail: string;
  displayName: string;
  action: string;
  licenceSku: string;
  beforeCost: number;
  afterCost: number;
  monthlySaving: number;
  annualisedSaving: number;
  approved: boolean;
  executed: boolean;
  executionMode: string;
  evidence: OutcomeLedgerEntryEvidence;
  approvedAt: string | null;
  executedAt: string | null;
  createdAt: string;
}

export interface ExecutionResult {
  recommendation: Recommendation;
  outcome: OutcomeLedgerEntry;
}

export interface OutcomesSummary {
  totalMonthlySaving: number;
  totalAnnualisedSaving: number;
  totalActionsExecuted: number;
  topPlaybook: string;
  avgMonthlySavingPerAction: number;
}

export interface RejectRecommendationBody {
  reason: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListRecommendationsParams = {
  status?: ListRecommendationsStatus;
  playbook?: string;
};

export type ListRecommendationsStatus =
  (typeof ListRecommendationsStatus)[keyof typeof ListRecommendationsStatus];

export const ListRecommendationsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  executed: "executed",
  all: "all",
} as const;

export type GenerateRecommendations200 = {
  generated: number;
  recommendations: Recommendation[];
};

export type ListOutcomesParams = {
  limit?: number;
};