# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Outcome Control Plane API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List all recommendations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, executed, all]
        - name: playbook
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
...
  /recommendations/{id}:
    get:
      operationId: getRecommendation
      tags: [recommendations]
      summary: Get a single recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /execution/approve/{id}:
    post:
      operationId: approveRecommendation
      tags: [execution]
      summary: Approve and execute a recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved and executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Cannot approve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /execution/reject/{id}:
    post:
      operationId: rejectRecommendation
      tags: [execution]
      summary: Reject a recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRecommendationBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Recommendation:
      type: object
      properties:
        id:
          type: integer
        userEmail:
          type: string
        displayName:
          type: string
        licenceSku:
          type: string
        monthlyCost:
          type: number
        annualisedCost:
          type: number
        trustScore:
          type: number
        executionStatus:
          type: string
          enum: [AUTO_EXECUTE, APPROVAL_REQUIRED, INVESTIGATE, BLOCKED]
        status:
          type: string
          enum: [pending, approved, rejected, executed]
        playbook:
          type: string
        connector:
          type: string
        lastActivity:
          type: string
          nullable: true
        daysSinceActivity:
          type: integer
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userEmail
        - displayName
        - licenceSku
        - monthlyCost
        - annualisedCost
        - trustScore
        - executionStatus
        - status
        - playbook
        - connector
        - lastActivity
        - daysSinceActivity
        - rejectionReason
        - createdAt
    ExecutionResult:
      type: object
      properties:
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        outcome:
          $ref: "#/components/schemas/OutcomeLedgerEntry"
      required: [recommendation, outcome]
    OutcomeLedgerEntry:
      type: object
      properties:
        id:
          type: integer
        recommendationId:
          type: integer
        userEmail:
          type: string
        displayName:
          type: string
        action:
          type: string
        licenceSku:
          type: string
        beforeCost:
          type: number
        afterCost:
          type: number
        monthlySaving:
          type: number
        annualisedSaving:
          type: number
        approved:
          type: boolean
        executed:
          type: boolean
        executionMode:
          type: string
        evidence:
          type: object
          additionalProperties: true
        approvedAt:
          type: string
          nullable: true
        executedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - recommendationId
        - userEmail
        - displayName
        - action
        - licenceSku
        - beforeCost
        - afterCost
        - monthlySaving
        - annualisedSaving
        - approved
        - executed
        - executionMode
        - evidence
        - approvedAt
        - executedAt
        - createdAt
...
    RejectRecommendationBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]